test(agent): move source-adjacent tests under tests
Keep map-phase and canopy coverage aligned with the repo's mirrored tests layout while preserving the existing schema and phase-loop coverage in their established test files.
